refactor(app): Sidebar.tsx 逻辑层拆六件
- 2847→1537 行：sidebarTypes 188 + sidebarSearch 63 + providerCatalogs 132 + threadProjections 310 + sessionFolderActions 829（含 projectionCache 整组）+ collapseAll 92
- JSX 渲染输出零改动（DOM/class 断言面不动）；renderWorkspaceEntry 按计划留后续单独 PR
- 循环依赖解开：showWorkspaceSessionMenu 回调留主文件；useThreadRows +9 additive 类型导出
- baseline 重生成 + 新文件登记豁免
验证：tsc 本族零 error；Sidebar 73 passed/1 failed 与基线同名（refresh 按钮既有）；gate 绿
（子代理 SidebarLogic 执行，主代理复核验证）
