+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T20:31:26.945Z",
+  "generatedAt": "2026-08-30T20:35:21.083Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
